added license headers, restructured package structure
