+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/error/ErrorView.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.hawkbit.ui;
+package org.eclipse.hawkbit.ui.error;
 
 import org.eclipse.hawkbit.ui.menu.DashboardMenu;
 import org.eclipse.hawkbit.ui.menu.DashboardMenuItem;
