+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/error/UiErrorDetails.java
@@ -1,4 +1,12 @@
-package org.eclipse.hawkbit.ui;
+/**
+ * Copyright (c) 2021 Bosch.IO GmbH and others.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+package org.eclipse.hawkbit.ui.error;
 
 public class UiErrorDetails {
     private static final UiErrorDetails EMPTY = new UiErrorDetails();
