+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/error/extractors/ConstraintViolationErrorExtractor.java
@@ -1,4 +1,12 @@
-package org.eclipse.hawkbit.ui;
+/**
+ * Copyright (c) 2021 Bosch.IO GmbH and others.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+package org.eclipse.hawkbit.ui.error.extractors;
 
 import java.util.Optional;
 import java.util.Set;
@@ -7,6 +15,7 @@ import java.util.stream.Collectors;
 import javax.validation.ConstraintViolation;
 import javax.validation.ConstraintViolationException;
 
+import org.eclipse.hawkbit.ui.error.UiErrorDetails;
 import org.eclipse.hawkbit.ui.utils.VaadinMessageSource;
 
 public class ConstraintViolationErrorExtractor extends AbstractSingleUiErrorDetailsExtractor {
